fixes and cleanups
git-svn-id: https://svn.apache.org/repos/asf/santuario/xml-security-java/trunk@1460408 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/xml/security/transforms/implementations/TransformXPath.java
@@ -147,7 +147,7 @@ public class TransformXPath extends TransformSpi {
                 Object[] eArgs = {currentNode};
                 throw new XMLSecurityRuntimeException("signature.Transform.node", eArgs, e);
             } catch (Exception e) {
-                Object[] eArgs = {currentNode, Short.valueOf(currentNode.getNodeType())};
+                Object[] eArgs = {currentNode, currentNode.getNodeType()};
                 throw new XMLSecurityRuntimeException("signature.Transform.nodeAndType",eArgs, e);
             }
         }
